Update: code refactored. different screen created
