+++ lib/main.dart
@@ -1,3 +1,4 @@
+import 'package:animation_graphics/screens/animated_builder.dart';
 import 'package:flutter/material.dart';
 import 'dart:math' show pi;
 //TODO: show?
@@ -6,35 +7,8 @@ void main() {
   runApp(const MainApp());
 }
 
-class MainApp extends StatefulWidget {
-  const MainApp({super.key});
-
-  @override
-  State<MainApp> createState() => _MainAppState();
-}
-
-class _MainAppState extends State<MainApp> with TickerProviderStateMixin {
-  late AnimationController _controller;
-  late Animation<double> _rotation;
-
-  @override
-  void initState() {
-    _controller = AnimationController(
-      vsync: this,
-      duration: const Duration(seconds: 5),
-    )..forward();
-
-    _rotation = Tween<double>(begin: 0, end: 2 * pi).animate(_controller);
-
-    _controller.repeat();
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
+class MainApp extends StatelessWidget {
+  const MainApp({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -42,39 +16,13 @@ class _MainAppState extends State<MainApp> with TickerProviderStateMixin {
       debugShowCheckedModeBanner: false,
       home: Scaffold(
         appBar: AppBar(
-          title: const Text("Animation"),
+          title: const Text("Animations"),
         ),
-        body: Center(
-          child: AnimatedBuilder(
-            animation: _controller,
-            builder: (BuildContext context, Widget? child) {
-              return Transform(
-                alignment: Alignment.center,
-                transform: Matrix4.identity()..rotateY(_rotation.value),
-                child: Container(
-                  height: 200,
-                  width: 200,
-                  decoration: BoxDecoration(
-                      color: Colors.purple,
-                      borderRadius: BorderRadius.circular(10.0),
-                      boxShadow: const [
-                        BoxShadow(
-                          color: Colors.black,
-                          blurRadius: 5.0,
-                          spreadRadius: 1.0,
-                          blurStyle: BlurStyle.solid,
-                          offset: Offset(1, 1),
-                        ),
-                      ]),
-                  child: const Center(
-                    child: Text("Animated Builder"),
-                  ),
-                ),
-              );
-            },
-          ),
-        ),
+        body: const AnimatedBuilderScreen(),
       ),
     );
   }
 }
+
+
+
