+++ lib/screens/animated_builder.dart
@@ -1,0 +1,68 @@
+import 'dart:math';
+
+import 'package:flutter/material.dart';
+
+class AnimatedBuilderScreen extends StatefulWidget {
+  const AnimatedBuilderScreen({super.key});
+
+  @override
+  State<AnimatedBuilderScreen> createState() => _AnimatedBuilderScreen();
+}
+
+class _AnimatedBuilderScreen extends State<AnimatedBuilderScreen> with TickerProviderStateMixin {
+  late AnimationController _controller;
+  late Animation<double> _rotation;
+
+  @override
+  void initState() {
+    _controller = AnimationController(
+      vsync: this,
+      duration: const Duration(seconds: 5),
+    )..forward();
+
+    _rotation = Tween<double>(begin: 0, end: 2 * pi).animate(_controller);
+
+    _controller.repeat();
+    super.initState();
+  }
+
+  @override
+  void dispose() {
+    _controller.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+          child: AnimatedBuilder(
+            animation: _controller,
+            builder: (BuildContext context, Widget? child) {
+              return Transform(
+                alignment: Alignment.center,
+                transform: Matrix4.identity()..rotateY(_rotation.value),
+                child: Container(
+                  height: 200,
+                  width: 200,
+                  decoration: BoxDecoration(
+                      color: Colors.purple,
+                      borderRadius: BorderRadius.circular(10.0),
+                      boxShadow: const [
+                        BoxShadow(
+                          color: Colors.black,
+                          blurRadius: 5.0,
+                          spreadRadius: 1.0,
+                          blurStyle: BlurStyle.solid,
+                          offset: Offset(1, 1),
+                        ),
+                      ]),
+                  child: const Center(
+                    child: Text("Animated Builder"),
+                  ),
+                ),
+              );
+            },
+          ),
+        );
+  }
+}
